[Trial]

+++ app/controllers/books_controller.rb
@@ -5,10 +5,15 @@ class BooksController < ApplicationController
   end
 
   def create
-    book = Book.new(book_params)
-    book.user_id = current_user.id
-    book.save
-    redirect_to book_path(book.id)
+    @book = Book.new(book_params)
+    if @book.save
+      redirect_to book_path(@book.id)
+      flash[:notice]="You have created book successfully."
+    else
+      @user = current_user
+      @books = Book.all
+      render :index
+    end
   end
 
   def show
@@ -24,14 +29,29 @@ class BooksController < ApplicationController
   end
 
   def edit
+    @book = Book.find(params[:id])
+    if @book.user == current_user
+      render :edit
+    else
+      redirect_to books_path
+    end
   end
 
   def update
+    @book = Book.find(params[:id])
+    if @book.update(book_params)
+      redirect_to book_path(@book.id)
+      flash[:notice]="You have updated book successfully."
+    else
+      render :edit
+    end
   end
 
   def destroy
     book = Book.find(params[:id])
-    book.destroy
+    if book.user_id == current_user.id
+      book.destroy
+    end
     redirect_to books_path
   end
